Propagate libvirt auth from actuator to libvirt client
In order to remove the use of qemu+tcp, we need a way to be able to
provide a ssh key for use with qemu+ssh, or TLS CA/certificate for
qemu+tls. This commit starts adding the plumbing for passing a SSH
private key, support for qemu+tls will be added later on.